requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pyodbc
python-jose[cryptography]
google-generativeai